Sort encounter orders by order number
This is an attempt to fix a bug which generates many failed events for
duplicate worklist item. I believe the bug was happening when a single
encounter had NEW order and corresponding DISCONTINUE orders in the same
encounter, but they were not in the correct order, so that DISCONTINUE
orders were failing to apply.

Other possible solutions could have been:
1. ignore NEW orders that have their dateStopped attribute set
2. when processing orders, go through the list and remove NEW -
DISCONTINUE order pairs (using previousOrderUuid)